AudioPolicyManager: Avoid invalid volume setting in APM
Do skip volume setting when invalid device id is sent as
argument to checkAndSetVolume().  If this setting is not skipped,
the audio volume index sent to checkAndSetVolume() may not match
with the actual audio device associated with outputDesc sent as
another argument, causing invalid volume set to actual audio device.

Change-Id: If2a38473480feea3fa14792de850d88a9a0fd835
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-71124
Signed-off-by: Aravind Siddappaji <aravindx.siddappaji@intel.com>