codec2: Add C2AllocatorBuf to implement a libdmabufheap BufferAllocator
Initial step to transition to libdmabufheap from ion.

This adds a C2AlllocatorBuf implementation which handles the
logic between allocating from DMABUF Heaps.

Also, to address the lack of !ION_FLAG_CACHED support in
the DMABUF Heaps (currently), in order to try to support
it for the ION fallback, I hope to use a "system-uncached"
heap (to be upstreamed into the kernel hopefully soon).

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I09ef198355140c28b50cc17b8a7acc54310e46fb